// gpj-client/src/shared/enums/trial-orderable-colum.ts
enum TrialOrderableColumn {
  TRIAL_ID = 'trialId',
  CLIENT = 'clientName',
  OFFICE = 'officeName',
  STATUS = 'orderableStatus',
  LAWYER = 'lawyer',
  EMAIL = 'email',
  TASK_TYPE = 'taskType',
  TRIAL_NUMBER = 'trialNumber',
  DEADLINE = 'deadline',
  SIDE = 'side',
  PLAINTIFF = 'plaintiff',
  DEFENDANT = 'defendant',
  PRIVATE_ANNOTATIONS = 'privateAnnotations',
  PUBLIC_ANNOTATIONS = 'publicAnnotations',
}

export default TrialOrderableColumn;